update modules using customed graph module
