--- Experiments/CustomizedDemo/STMeta_v0.model.yml
@@ -1,24 +1,0 @@
-# network structure parameters
-st_method: 'LSTM'
-temporal_merge: 'gal'
-graph_merge: 'gal'
-
-# gcn parameters
-gcn_k: 0
-gcn_layers: 1
-gclstm_layers: 1
-
-# LSTM units
-num_hidden_units: 64
-# dense units
-num_filter_conv1x1: 32
-
-build_transfer: False
-
-# merge parameters
-graph_merge_gal_units: 64
-graph_merge_gal_num_heads: 2
-temporal_merge_gal_units: 64
-temporal_merge_gal_num_heads: 2
-
-model_version: 'TMeta'
