--- Experiments/CustomizedDemo/metro_shanghai.data.yml
@@ -1,29 +1,0 @@
-# dataset and city
-dataset: Metro
-city: Shanghai
-
-closeness_len: 6
-period_len: 7
-trend_len: 4
-
-graph: Distance-Correlation-Line
-
-data_range: all
-train_data_length: all
-
-threshold_distance: 5000
-threshold_correlation: 0.7
-threshold_interaction: 30
-threshold_neighbour: 23
-
-normalize: True
-train: True
-
-lr: 1e-5
-early_stop_length: 200
-early_stop_patience: 0.1
-max_epoch: 20000
-batch_size: 64
-
-group: Shanghai
-mark: BM
