+++ Experiments/ST_MGCN/ST_MGCN_Obj.py
@@ -40,7 +40,7 @@ def stmeta_param_parser():
     # device parameter
     parser.add_argument('--Device', default='0', type=str)
     # version control
-    parser.add_argument('--CodeVersion', default='V0')
+    parser.add_argument('--CodeVersion', default='V')
     # Merge times
     parser.add_argument('--MergeIndex', default=6, type=int)
     return parser
@@ -70,18 +70,33 @@ else:
     else:
         current_device = str(deviceIDs[0])
 
+
+
 # Config data loader
 data_loader = NodeTrafficLoader(dataset=args['Dataset'], city=args['City'],
                                 data_range=args['DataRange'], train_data_length=args['TrainDays'],
                                 closeness_len=int(args['CT']), period_len=int(args['PT']), trend_len=int(args['TT']),
-                                threshold_interaction=args['TI'], threshold_distance=args['TD'],
-                                threshold_correlation=args['TC'], graph=args['Graph'], with_lm=True, normalize=True, MergeIndex=args['MergeIndex'],
+                                normalize=True, MergeIndex=args['MergeIndex'],
                                 MergeWay="max" if args["Dataset"] == "ChargeStation" else "sum")
+
+# Import the Class:GraphGenerator
+from UCTB.preprocess.CusGraph import CusGraph
+# Call GraphGenerator to initialize and generate LM
+graph = args['Graph']
+graphBuilder = CusGraph(graph,
+                             dataset=data_loader.dataset,
+                             train_data = data_loader.train_data,
+                             traffic_data_index = data_loader.traffic_data_index,
+                             train_test_ratio = data_loader.train_test_ratio,
+                             threshold_distance=args['TD'],
+                             threshold_correlation=args['TC'],
+                             threshold_interaction=args['TI'],
+                             )
 
 ST_MGCN_Obj = ST_MGCN(T=int(args['CT']) + int(args['PT']) + int(args['TT']),
                       input_dim=1,
                       external_dim=data_loader.external_dim,
-                      num_graph=data_loader.LM.shape[0],
+                      num_graph=graphBuilder.LM.shape[0],
                       gcl_k=args['K'],
                       gcl_l=args['L'],
                       lstm_units=args['LSTMUnits'],
@@ -101,7 +116,7 @@ if args['Train'] == 'True':
     ST_MGCN_Obj.fit(traffic_flow=np.concatenate((np.transpose(data_loader.train_closeness, [0, 2, 1, 3]),
                                                  np.transpose(data_loader.train_period, [0, 2, 1, 3]),
                                                  np.transpose(data_loader.train_trend, [0, 2, 1, 3])), axis=1),
-                    laplace_matrix=data_loader.LM,
+                    laplace_matrix=graphBuilder.LM,
                     target=data_loader.train_y,
                     external_feature= None,
                     early_stop_method='t-test',
@@ -119,7 +134,7 @@ ST_MGCN_Obj.load(code_version)
 prediction = ST_MGCN_Obj.predict(traffic_flow=np.concatenate((np.transpose(data_loader.test_closeness, [0, 2, 1, 3]),
                                                              np.transpose(data_loader.test_period, [0, 2, 1, 3]),
                                                              np.transpose(data_loader.test_trend, [0, 2, 1, 3])), axis=1),
-                                 laplace_matrix=data_loader.LM,
+                                 laplace_matrix=graphBuilder.LM,
                                  external_feature=None,
                                  sequence_length=data_loader.test_sequence_len,
                                  output_names=['prediction'],
